requirements.txt
python-dotenv
requests
msal
boto3
PyPDF2
